# docs/openapi.yaml
openapi: 3.1.0
info:
  title: EIOS and LMS Integration API Contract
  description: >
    This API contract specifies the shared integration endpoints, request/response shapes,
    and webhook payloads for LMS metadata, role synchronization, and notification subscriptions
    between the EIOS Knowledge Base and Moodle LMS systems.

    Roles supported inside EIOS:
    - ADMINISTRATOR
    - CONTENT_MANAGER
    - TEACHER
    - LEARNER
  version: 1.0.0
  license:
    name: Proprietary
    url: "https://example.com/license"
servers:
  - url: /api/v1/integration
    description: Default integration base URL
security:
  - bearerAuth: []
paths:
...
  /tasks/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Get a task by ID
      operationId: getTaskById
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Task not found
        '500':
          $ref: '#/components/responses/InternalServerError'
  /tasks/{id}/transition:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    post:
      summary: Transition a task state
      description: Transitions a task state atomically with a conditional/optimistic concurrency check.
      operationId: transitionTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - expectedStatus
                - newStatus
              properties:
                expectedStatus:
                  type: string
                newStatus:
                  type: string
      responses:
        '200':
          description: Task successfully transitioned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Task not found
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Task:
      type: object
      required:
        - id
        - title
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: Unique string identifier for the task (UUID)
        title:
          type: string
          description: Task title
        status:
          type: string
          description: Task status (e.g. pending_review, resolved)
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required:
        - errorClass
        - errorCode
        - message
        - timestamp
      properties:
        errorClass:
          type: string
          description: Categorization of the error (e.g. VALIDATION_ERROR, INTEGRATION_ERROR, SECURITY_ERROR)
        errorCode:
          type: string
          description: Precise internal identifier of the specific defect class
        message:
          type: string
          description: Human-readable explanation of the defect context
        timestamp:
          type: string
          format: date-time
          description: ISO-8601 timestamp representing when the error was registered
  responses:
    BadRequestError:
      description: Invalid request payload, query parameters, or formatting
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnauthorizedError:
      description: Security token missing, malformed, or expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: An unhandled exception or upstream system failure occurred
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'